requirements.txt
# Core runtime dependencies for LongVideoAI
moviepy==2.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
numpy==2.0.2
pillow==11.3.0
python-dotenv==1.1.1
requests>=2.31.0
PyYAML>=6.0.2
